+++ store/snapshots/manager_test.go
@@ -35,6 +35,8 @@ func TestManager_List(t *testing.T) {
 	list, err := manager.List()
 	require.NoError(t, err)
 	assert.Equal(t, []*types.Snapshot{}, list)
+
+	manager.Close() // check api call compiles - does not return any value
 }
 
 func TestManager_LoadChunk(t *testing.T) {